refactor(ui): split AnalyzePane into cohesive hooks and subcomponents
